requirements.txt
# Core ML Libraries
tensorflow==2.15.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Data Collection
yfinance==0.2.28
requests==2.31.0

# Visualization
matplotlib==3.7.2
plotly==5.15.0
seaborn==0.12.2

# Web Application
streamlit==1.25.0

# Technical Indicators
ta==0.10.2
pandas-ta==0.3.14b0

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2
joblib==1.3.2

# Progress bars
tqdm==4.65.0
